mirroring: async push to slaves...
tested by a few people independently, on ML and on #gitolite

Note that the changes to the (private) test suite use an arbitrary "sleep 5"
to allow the other sides to settle.  Without this, there were too many errors
because the test suite was trying the next command before the previous one had
effectively finished.