redis caching: please read below
This commit brings in redis caching to Gitolite.

Documentation is not yet done, but briefly:
1.  make sure you have Redis available ('perl -MRedis -e 0' should succeed)
2.  enable caching by uncommenting the "CACHE => 'redis'" line in the rc file

However, this code is not the same as the old "redis" branch (which has now
been deleted).  It is a complete rewrite, much simpler, and much more elegant.
It uses the ideas of memoisation rather than explicit "cache_set" and
"cache_get", and at present memoizes just one function: the access() function
in Load.pm.  That seems to be sufficient for most purposes anyway.