ssh-authkeys: update authkeys when keydir is present but empty
When ~/.gitolite/keydir is present but empty, and someone runs 'gitolite
trigger POST_COMPILE', gitolite does not clear out the corresponding lines in
~/.ssh/authorized_keys.

----

NOTE that this can only happen when you're managing gitolite directly on the
server, per http://gitolite.com/gitolite/odds-and-ends.html#server-side-admin.

When you're managing gitolite remotely, and you 'git rm -rf keydir' in your
admin clone, then commit and push, this behaviour (of not updating the
authkeys file in ~/.ssh) will remain.

However, it's not by design -- it's more of a lucky accident!  What's
happening here is, when ~/.gitolite is updated via a "git checkout -f", the
(now empty) 'keydir' directory is *completely deleted*.  This makes
ssh-authkeys bail much earlier, because now it looks like a system that is not
even *using* ssh keys (for example, an http-mode installation).

The end result is the same -- the authkeys file is left untouched.  However, I
will not be "fixing" that, for obvious reasons.