allow repo symlinks while still being paranoid...
I don't like "find -L" in a security program.  Period.  If you MUST keep some
of your repos somewhere else, you'll need to do it this way.

Let's say all the repos under foo/bar and baz are actually on a different
disk.  That is, ~/repositories/foo/bar is a symlink to some other disk, as is
~/repositories/baz.

Then add this to the rc file somewhere in the "rc variables used by various
features" section:

    REPO_SYMLINKS   =>  "foo/bar/ baz/",

DON'T forget the trailing slash at the end of each name, and the comma
outside.