create command: remove race condition...
Because of the pipe in the create command, 'gitolite git-config' and 'perms
-c' would run simultaneously.  The problem was, if git-config ran between
these two statements in new_wild_repo() in Store.pm:

    new_repo($repo);
        # 'gitolite git-config' runs here
    _print( "$repo.git/gl-creator", $user );

*and* the repo pattern had CREATOR in it, it would not return the correct
values for the default.roles options, since the repo pattern would not be
deemed to match the actual repo (between creator() and generic_name()).

Thanks to Ronald Wirth for catching this; see [1] for details.

[1]: https://groups.google.com/forum/#!topic/gitolite/5Dv6ViDmfF4

----

...and while we're about it, I changed it to use /bin/sh and smoke tested it
using "dash", which I believe is close enough.