fix ref-create permission bug on wild repos
TLDR: users are able to *create* new refs that they are not supposed to.
Upgrading gitolite is not mandatory; there is a workaround within the conf
syntax itself (see below).

The bug shows up if the following four conditions are satisfied:

    repo foo/..*                    # 1 (see below)
        C       =   @all            # 2
        RW+CD   =   CREATOR         # 3
        RW      =   special_user    # 4

1.  must be a wild repo
2.  the create-repo line must allow "special_user(s)"
3.  there is at least one rule containing RW.*C in the repo
4.  the intent is that special user(s) can ff-push but cannot *create* refs
    since they have only RW; for a refresher on this, see
    http://gitolite.com/gitolite/conf.html#write-types

In such cases, the restriction on creating a ref is ignored.  (Only creation
is affected; the bug does not affect delete, rewind permissions).

In general, this is a relatively minor bug.  No data is destroyed, and no data
is leaked.  Some cleanup of useless refs may be required if someone has
(ab)used this.  At worst this could be a potential DOS, but I have never
considered DOS to be a valid concern when it requires *authorised* users.

If you have such a conf, and cannot immediately upgrade, there is a
workaround:

    repo foo/..*
        RW+CD   =   CREATOR
        RW      =   special_user

        -       =   special_user    # add a deny rule for the special user(s)
        C       =   @all            # move the repo-create perm to the end,
                                    # *after* the deny rule above

(Thanks to hseg on #gitolite for catching this and being tenacious about it!
At first, I was arrogant enough to reject the idea that such a bug could
exist; it took me a bit to get the right conditions in place and see the
problem!)