new mirror function: 'status all all'
We used to say if you need the status of all slaves for all repos you have to
roll it yourself, maybe like this:

    gitolite list-phy-repos | while read r
    do
        echo ---- $r
        gitolite mirror list slaves $r
    done

This isn't great for automation.

The new feature simply prints a list of repos that have *some* error, which is
arguably more useful for further action/processing.