ssh-authkeys-split: avoid creating invalid keyfiles
Verify that each line from a multiline keyfile is plausible using
`ssh-keygen -l` to generate a fingerprint. This is similar to the
check performed by the main ssh-authkeys script, except we don't
bother checking the fingerprint format in ssh-authkeys-split.

This should reduce the damage due to problems such as stray blank
lines or unexpected key formats (e.g. PuTTY keys).