allow creator check to be bypassed during mirroring...
The original, intended, purpose of wild repos is that they belong to the user,
and if the user is gone, so should his wild repos.

However, it seems people are using the wild repos thing as a convenience to
avoid creating actual repo stanzas in the conf file.  For them, the actual
creator of a repo is more like the gitolite "admin" - his *authority* is being
to used to create something, but the thing that is created is not tied to his
*identity*.

Oh well... so be it!

To use, just add

    option bypass-creator-check = 1

to the rules for the repo.