Add support for Github's new TEMPLATE features
Github recently added support for TEMPLATEs for certain Github-specific
activities, including creating Issues and Pull Requests. This patch
creates these files as symlinks to the CONTRIBUTING document, which
explains the process that should be used.

Signed-off-by: Eugene E. Kashpureff Jr <eugene@kashpureff.org>

Committer's note: I still refuse to use pull requests that *require me to go
to the website and do stuff there*.  But it seems it's easy enough if the
requestor gives you a number.  In this case, it was

    git fetch github refs/pull/78/head
    git merge FETCH_HEAD