suppress warnings when trying to clean repo hook
suppress warning when setting the repo-specific hooks to an empty
string, e.g. removing an existing hook.

Specifically, we default repositories to use a repo-specific hook:

repo @all
    option hook.post-update = jenkins

this causes gitolite to output a warning that it is ignoring the attempt
to set a post-update hook on the admin repository.    I'd like to be
able to do:

repo gitolite-admin
    option hook.post-update = ""

but the warning is printed before it realizes that the argument is empty.
This patch just changes the order of the checks so that the warning is
only printed if the hooks are actually defined.