ssh-authkeys-split: do not zap keys without trailing newlines
The normal ssh-authkeys script handles key files without trailing newlines
OK. However ssh-authkeys-split does not, because it tries to treat a
zero-newline file as a multi-line file, and the shell `read` builtin will
not read a line that ends with eof rather than a newline. So it ends up
discarding zero-newline keys.