reduce stat() and other calls for huge installs
Fedora has 42000 or so repos, and a 'gitolite compile' was taking too
long.

This set of changes reduces the number of stat() and other calls from 9
to 2 per unchanged, existing repo.  (The expectation is that only a few
repos' rules are being changed each time, so this helps somewhat
optimise the others not to take the same amount of time).