Reduce disk writes in post-compile triggers
We have a system with thousands of repositories located on disks where:

1. write operations are much slower than read operations
2. IO is burstable but is throttled after a certain limit

In such conditions, writing out 10-15 git config settings per each of the
15,000 repositories resulted in around 200,000 calls to git-config and
200,000 disk writes to 15,000 files on every gitolite-admin push, even
if there were no changes to any git configs.

With this change, we call git-config once to read the entirety of the
git config for the repo, and issue writes only if there are changes to
any of the settings.

Similarly, we only touch git-daemon-export-ok if such file doesn't
already exist to avoid another few thousand unnecessary writes.

As a result, our gitolite-admin push times went from 10-12 minutes to 20
seconds.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>