'info' learns '-p' option to show only physical repos
While we were meaninglessly debating (on the mailing list) the merits of
showing the wild repo regexes to a *human*, someone came up with a
reasonable need for an option to *not* show them so that certain
automated tasks can be a wee bit simpler.  Note that I still think the
regexes *should* be shown to the user by default; we're just adding an
option to explicitly suppress them.

15:31:59  <hiroki> sitaram: Thanks ! The most common use case would be
    for a sanity check script running on slaves that will check what
    repositories it has locally, and compare it with the output of info
    to their master. That way they can be certain all repositories are
    synced over properly. Second step would be to check all commit
    hashes of all repositories locally and their respective master.