access: fallthru in VREFs needs fixing
Björn Kautler pointed out that for VREFs suffering fallthru, the
interpretation of the result of access would be wrong.

Initially, I was reluctant to make that adjustment here.  The 'access'
command is meant to be just a shell entry point for the access()
function, nothing more, so bringing in adjustments that occur elsewhere
in gitolite would not be right.

But it makes automated testing for VREF rules more difficult (and VREFs
are already hard enough for most people to understand), so I caved.

Hopefully this won't turn into a slippery slope...