block access() to repos being imported
dc13dfc introduced a block on accessing repos which are in the process
of being moved into gitolite's control.

This block was (a) in gitolite-shell, which would catch all git-client
activity and (b) in list_phy_repos(), which would prevent those repos
from being seen by the 'info' command.

Unfortunately that was stupid; it also blocked 'gitolite setup' itself,
because setup uses list_phy_repos!

The correct place to put this was always going to be access(), but I had
initially shied away from that because it would cause a slight glitch in
the working of any POST_COMPILE trigger scripts that used the access()
function on any of the newly migrated repos.

But nothing else really works.  As a result, the step where you run
`gitolite setup` when importing now becomes:

    gitolite compile
    gitolite setup --hooks-only
    gitolite trigger POST_COMPILE