repo-specific hooks: fix bug in handling reserved hooks
There are two points to consider here:
-   is the user trying to delete all hooks for a repo?
-   is it a reserved hook (one we should not be touching)

The second one is more important; the first one is a spurious warning
that is at worst an aesthetic problem.

Unfortunately, when refactoring to fix a bug in 7898f9b, I gave higher
priority to the wrong issue, and the more important one was not checked
when the less important one was "true".

As a result, control moved to the second stage, where hooks were
symlinked to the multi-hook driver.  Including the all important
'post-update' hook in 'gitolite-admin'.

Ouch!