memberships: don't try to match patterns for users!
Technically, the previous behaviour could even be called a "feature", in
that it allowed you to use regexes for users -- contradicting the
documentation! -- in the following way:

    @foo = u[123] u9
    repo bar
        RW+ =   @foo

which would enable users u1, u2, and u3 to access bar.

I thought for a bit about this, and decided to fix the code rather than
the documentation.

Leaving it as a quirk was another alternative, but it would only be a
quirk [1], and an incomplete one at that because it only works for
patterns inside *groups*, not patterns that are actually *on* the rule
line (e.g. RW+ = u[123]).  No good...

Thanks to Steven Peckins for catching this.

[1]:

    $ dict quirk
    From The Collaborative Fictional Dictionary of English v.0.48 [gcide]:
        Quirk \Quirk\ (kw[~e]rk), n.  [Written also {querk}.]
        1.  a bug that doesn't do enough harm to be shot down with extreme
            prejudice, and in fact might even be said to do something
            vaguely OK.  If you squint.  And if it's a Sunday.