'config' user command needs to allow for space-separated config values
This command:

    ssh git@host config my/wild/repo --add gitweb.owner "Sitaram Chamarty"

doesn't work.  Sshd converts it to

    config my/wild/repo --add gitweb.owner Sitaram Chamarty

so by the time src/commands/config sees this, that looks like two
"values" ("Sitaram" and "Chamarty") instead of one.

We can't do much about this in the general case; that's how sshd rolls.
I suppose we could, in parse_soc() in src/gitolite-shell, use something
like Text::ParseWords instead of a simple "split" (and then send in
arguments with backslash-escaped spaces etc) but that's too big a change
impacting too many things.  Not going to happen for this simple need.

The simplest way out is for src/commands/config to learn that multiple
trailing arguments are actually one space-separated argument.