When prepending our options to a key, check if key file already has any options.  In this case put comma after end of our options instead of space.
To tell key with options against bare key we check if key starts
with (ecdsa|ssh)-.  At the moment of this commit OpenSSH supports
the following algorithms “ecdsa-sha2-nistp256”, “ecdsa-sha2-nistp384”,
“ecdsa-sha2-nistp521”, “ssh-ed25519”, “ssh-dss” or “ssh-rsa”.  If
a different algorithm to be introduce to OpenSSH in future, this
code would be broken.  However, adding a full key parser to this
subroutine seems to be overkill now.