gitolite mirroring terminology changes
This affects the mirroring code and documentation:
"slave"/"slaves" are now "copy"/"copies".

Backward compatibility should be maintained; you do not need to
change either your gitolite.conf, or any scripts you have
written on top, until you are ready to do so.  (This in turn
means the word "slave" will still be present in the code, though
only just as much as needed.)

Should you wish to make this change, you need to migrate to the
latest version (which is also tagged as 3.6.12, so if you want
to wait till the distros pick it up wait for that), and then:

-   In the gitolite.conf file, change `option mirror.slaves` to
    `option mirror.copies`.

-   If you have any scripts that use the `gitolite mirror list
    slaves` command, change to `gitolite mirror list copies`.

sitaram