Take public keys at newline, instead of on EOF
This is better UX: many people these days are unaware of what the EOF character is, and instead expect programs to react on pressing enter.

`sskm` and the docs (https://gitolite.com/gitolite/contrib/sskm.html#step-1-adding-the-key) show adding a key by piping in. However in my experience actually deploying gitolite, people are often more comfortable pasting a key in: it is easier to cat + copy + paste than to scp. But with `safe_stdin` these people are left hanging, totally at a loss of how to administer the server or administer their own keys. To support these people, I have patched my deployment with this, and it has already saved a lot of tech support time.

I realize that 'safe_stdin' has 'safe' in its name for a reason. I hope you will consider the tradeoff.